+++ lib/features/notifications/domain/usecase/mark_as_read_all_notifications_use_case.dart
@@ -1,0 +1,14 @@
+import 'package:zachranobed/features/notifications/domain/repository/notifications_repository.dart';
+import 'package:zachranobed/models/user_data.dart';
+
+/// A use case to mark all unread notifications as read for the given [user].
+class MarkAsReadAllNotificationsUseCase {
+  final NotificationsRepository _repository;
+
+  MarkAsReadAllNotificationsUseCase(this._repository);
+
+  /// Marks all unread notifications as read for the given [user].
+  Future<void> invoke(UserData user) {
+    return _repository.markAllAsRead(user: user);
+  }
+}