+++ lib/services/di/services_dependency_container.dart
@@ -5,6 +5,7 @@ import 'package:zachranobed/services/auth_service.dart';
 import 'package:zachranobed/services/carrier_service.dart';
 import 'package:zachranobed/services/configuration_service.dart';
 import 'package:zachranobed/services/delivery_service.dart';
+import 'package:zachranobed/services/entity_notification_service.dart';
 import 'package:zachranobed/services/entity_pairs_service.dart';
 import 'package:zachranobed/services/entity_service.dart';
 import 'package:zachranobed/services/food_box_service.dart';
@@ -17,6 +18,7 @@ class ServicesDependencyContainer {
     GetIt.I.registerSingleton(FoodBoxService());
     GetIt.I.registerSingleton(EntityService());
     GetIt.I.registerSingleton(EntityPairService());
+    GetIt.I.registerSingleton(EntityNotificationService());
     GetIt.I.registerSingleton(MealService());
     GetIt.I.registerSingleton(DeliveryService());
     GetIt.I.registerSingleton(